+++ src/weather/provider.rs
@@ -1,12 +1,16 @@
-use super::{GetWeather, OpenMeteo, OpenWeatherMap, RustormyError, Weather};
+use super::{GetWeather, RustormyError, Weather};
 use crate::config::Config;
 use crate::models::{Location, Provider};
+use crate::weather::open_meteo::OpenMeteo;
+use crate::weather::open_weather_map::OpenWeatherMap;
+use crate::weather::world_weather_online::WorldWeatherOnline;
 use enum_dispatch::enum_dispatch;
 
 #[enum_dispatch(GetWeather)]
 pub enum GetWeatherProvider {
     OpenMeteo,
     OpenWeatherMap,
+    WorldWeatherOnline,
 }
 
 impl From<Provider> for GetWeatherProvider {
@@ -14,6 +18,7 @@ impl From<Provider> for GetWeatherProvider {
         match provider {
             Provider::OpenMeteo => OpenMeteo::default().into(),
             Provider::OpenWeatherMap => OpenWeatherMap::default().into(),
+            Provider::WorldWeatherOnline => WorldWeatherOnline::default().into(),
         }
     }
 }